Show last saved file/path & prevent file overwrite
Show last saved file next to the Start Index input in order to help people see which number to start their generation at.

Also add some more error handling for the SaveAs funcs - now we will warn the user when they are overwriting a file and prevent that from happening.